+++ builder/setup-builder.nix
@@ -5,10 +5,12 @@
 , prePatch ? null, postPatch ? null
 , preBuild ? component.preBuild , postBuild ? component.postBuild
 , preInstall ? component.preInstall , postInstall ? component.postInstall
-, cleanSrc ? haskellLib.cleanCabalComponent package component src
+, cleanSrc ? haskellLib.cleanCabalComponent package component "setup" src
 }:
 
 let
+  cleanSrc' = haskellLib.rootAndSubDir cleanSrc;
+
   fullName = "${name}-setup";
 
   includeGhcPackage = lib.any (p: p.identifier.name == "ghc") component.depends;
@@ -35,7 +37,7 @@ let
   drv =
     stdenv.mkDerivation ({
       name = "${ghc.targetPrefix}${fullName}";
-      src = cleanSrc;
+      src = cleanSrc'.root;
       buildInputs = component.libs
         ++ component.frameworks
         ++ builtins.concatLists component.pkgconfig;
@@ -44,7 +46,10 @@ let
       passthru = {
         inherit (package) identifier;
         config = component;
-        inherit configFiles cleanSrc;
+        srcSubDir = cleanSrc'.subDir;
+        srcSubDirPath = cleanSrc'.root + cleanSrc'.subDir;
+        cleanSrc = cleanSrc';
+        inherit configFiles;
       };
 
       meta = {
@@ -79,6 +84,13 @@ let
         runHook postInstall
       '';
     }
+    // (lib.optionalAttrs (cleanSrc'.subDir != "") {
+      prePatch =
+        # If the package is in a sub directory `cd` there first
+        ''
+          cd ${lib.removePrefix "/" cleanSrc'.subDir}
+        '';
+    })
     // (lib.optionalAttrs (patches != []) { patches = map (p: if builtins.isFunction p then p { inherit (package.identifier) version; inherit revision; } else p) patches; })
     // hooks
   );