+++ lib/ghcjs-project.nix
@@ -35,7 +35,16 @@
     version = "3.2.5";
     materialized = ../materialized/ghcjs/alex + "/${compiler-nix-name}";
   }
-, cabal-install ? pkgs.haskell-nix.cabal-install.${compiler-nix-name}
+, cabal-install ? pkgs.haskell-nix.tool compiler-nix-name "cabal" {
+    index-state = pkgs.haskell-nix.internalHackageIndexState;
+    version = "3.2.0.0";
+    # Cabal 3.2.1.0 no longer supports he mix of `cabal-version`,
+    # lack of `custom-setup` and `v1-install` used by ghcjs boot.
+    cabalProjectLocal = ''
+      constraints: Cabal <3.2.1.0
+    '';
+    materialized = ../materialized/ghcjs/cabal + "/${compiler-nix-name}";
+  }
 , ...
 }@args:
 let
@@ -129,6 +138,7 @@ let
             (n: _: !(builtins.any (x: x == n)
                 ["src" "ghcjsVersion" "ghcVersion" "happy" "alex" "cabal-install"])) args) // {
         src = configured-src;
+        index-state = "2020-11-16T00:00:00Z";
         compiler-nix-name = if isGhcjs88 then "ghc884" else "ghc865";
         configureArgs = pkgs.lib.optionalString isGhcjs88 "--constraint='Cabal >=3.0.2.0 && <3.1'";
         materialized = ../materialized + (if isGhcjs88 then "/ghcjs884" else "/ghcjs865");