+++ lib/import-and-filter-project.nix
@@ -37,11 +37,18 @@ in project // {
                       then
                         pkgs.lib.lists.elemAt sourceRepos (
                           pkgs.lib.strings.toInt (pkgs.lib.strings.removePrefix srcRepoPrefix subDir))
-                      else if haskellLib.canCleanSource srcRoot
-                        then haskellLib.cleanSourceWith { src = srcRoot; inherit subDir; }
-                        else srcRoot + (if subDir == "" then "" else "/" + subDir);
+                      else haskellLib.appendSubDir {
+                          src = srcRoot;
+                          inherit subDir;
+                          includeSiblings = true; # Filtering sibling dirs of the package dir is done in the
+                                                  # component builder so that relative paths can be used to
+                                                  # reference project directories not in the package subDir.
+                        };
             in oldPkg // {
               src = (pkgs.lib).mkDefault packageSrc;
+              package = oldPkg.package // {
+                isProject = (pkgs.lib).mkDefault true;
+              };
             }) old;
     };
   }