+++ modules/configuration-nix.nix
@@ -8,4 +8,11 @@
   # this if we used the terminfo shipped with GHC, but this package is
   # reinstallable so we'd rather have it defined in the plan.
   packages.terminfo.components.library.libs = [pkgs.ncurses];
+
+  # The `extra-libraries` field in `X11.cabal` does not include Xss and Xinerama
+  # see https://github.com/input-output-hk/haskell.nix/pull/988
+  packages.X11.components.library.libs = [
+    pkgs.xorg.libXScrnSaver
+    pkgs.xorg.libXinerama
+  ];
 }