+++ overlays/default.nix
@@ -10,6 +10,7 @@ let
     bootstrap = import ./bootstrap.nix;
     ghc = import ./ghc.nix;
     ghc-packages = import ./ghc-packages.nix;
+    darwin = import ./darwin.nix;
     windows = import ./windows.nix;
     armv6l-linux = import ./armv6l-linux.nix;
     musl = import ./musl.nix;
@@ -17,7 +18,6 @@ let
     emscripten = import ./emscripten.nix;
     nix-prefetch-git-minimal = import ./nix-prefetch-git-minimal.nix;
     gobject-introspection = import ./gobject-introspection.nix;
-    shims = import ./shims.nix;
     eval-on-current = import ./eval-on-current.nix;
     eval-on-build = import ./eval-on-build.nix;
     ghcjs = import ./ghcjs.nix;
@@ -43,6 +43,7 @@ let
     bootstrap
     ghc
     ghc-packages
+    darwin
     windows
     armv6l-linux
     musl
@@ -51,7 +52,6 @@ let
     nix-prefetch-git-minimal
     ghcjs
     gobject-introspection
-    shims
     # Restore nixpkgs haskell and haskellPackages
     (_: prev: { inherit (prev.haskell-nix-prev) haskell haskellPackages; })
   ];