+++ test/shell-for/default.nix
@@ -24,6 +24,8 @@ let
     # they use a nix-shell --pure. Normally you would BYO cabal-install.
     tools = { cabal = "3.2.0.0"; };
     exactDeps = true;
+    # Avoid duplicate package issues when runghc looks for packages
+    packageSetupDeps = false;
   };
 
   envPkga = pkgSet.config.hsPkgs.shellFor {
@@ -33,6 +35,8 @@ let
     # they use a nix-shell --pure. Normally you would BYO cabal-install.
     tools = { cabal = "3.2.0.0"; };
     exactDeps = true;
+    # Avoid duplicate package issues when runghc looks for packages
+    packageSetupDeps = false;
   };
 
   envDefault = pkgSet.config.hsPkgs.shellFor {
@@ -42,6 +46,8 @@ let
     # This adds cabal-install to the shell, which helps tests because
     # they use a nix-shell --pure. Normally you would BYO cabal-install.
     tools = { cabal = "3.2.0.0"; };
+    # Avoid duplicate package issues when runghc looks for packages
+    packageSetupDeps = false;
   };
 
 in recurseIntoAttrs {
@@ -64,8 +70,10 @@ in recurseIntoAttrs {
       touch $out
     '';
 
-    meta.platforms = platforms.all;
-    meta.disabled = stdenv.hostPlatform.isMusl || stdenv.hostPlatform.isWindows;
+    meta = {
+      platforms = platforms.unix;
+      disabled = stdenv.hostPlatform.isMusl;
+    };
 
     passthru = {
       # Used for debugging with nix repl